LayoutTests/fast/security-labels/navigate-event-expected.txt

CONSOLE MESSAGE: Not allowed to load resource by checknavigate: http://localhost/bar
CONSOLE MESSAGE: Not allowed to load resource by checknavigate: http://localhost/foo
CONSOLE MESSAGE: Not allowed to load resource by checknavigate: http://localhost/bar
CONSOLE MESSAGE: Not allowed to load resource by checknavigate: http://localhost/foo
This tests that the checknavigate event behaves properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


checknavigate reached with http://localhost/bar
PASS tag.isOn(event.securityLabel) is false
checknavigate reached with http://localhost/foo
PASS tag.isOn(event.securityLabel) is false
checknavigate reached with http://localhost/bar
PASS tag.isOn(event.securityLabel) is true
checknavigate reached with http://localhost/foo
PASS tag.isOn(event.securityLabel) is true
PASS successfullyParsed is true

TEST COMPLETE
